refactor: split retrieval portfolio jobs as discovery
