+++ demo/src/main/java/com/example/demo/portfolio/controller/PortfolioController.java
@@ -2,8 +2,6 @@ package com.example.demo.portfolio.controller;
 
 import com.example.demo.member.service.CustomUserDetailsService;
 import com.example.demo.portfolio.dto.PortfolioDTO;
-import com.example.demo.portfolio.dto.PortfolioOverviewDTO;
-import com.example.demo.portfolio.dto.PortfolioSearchDTO;
 import com.example.demo.portfolio.service.PortfolioSearchService;
 import com.example.demo.portfolio.service.PortfolioService;
 import com.example.demo.review.dto.ReviewDTO;
@@ -82,14 +80,6 @@ public class PortfolioController {
         log.info("Deleted portfolio with ID: {}", portfolioId);
         return ResponseEntity.ok().build();
     }
-//
-//    @GetMapping("/shared/soft-deleted")
-//    @Operation(summary = "[공통] soft-deleted가 된 포트폴리오 조회")
-//    public ResponseEntity<List<PortfolioOverviewDTO.Response>> getAllSoftDeleted() {
-//        List<PortfolioOverviewDTO.Response> softDeletedPortfolios = portfolioService.getAllSoftDeletedPortfolios();
-//        log.info("Fetched all soft-deleted portfolios");
-//        return ResponseEntity.ok(softDeletedPortfolios);
-//    }
 
     @GetMapping("/shared/all")
     @Operation(summary = "[공통] soft-deleted 를 포함한 전체 포트폴리오 조회")
@@ -99,22 +89,6 @@ public class PortfolioController {
         return ResponseEntity.ok(portfolioResponses);
     }
 
-    @GetMapping("/shared/search")
-    @Operation(summary = "[공통] 포트폴리오 검색하여 조회")
-    public ResponseEntity<List<PortfolioSearchDTO.Response>> getSearchPortfolio(
-            @Parameter(description = "검색 키워드")
-            @RequestParam String content) {
-        List<PortfolioSearchDTO.Response> searchResult = portfolioSearchService.search(content);
-        log.info("Searched portfolios with content: {}", content);
-        return ResponseEntity.ok(searchResult);
-    }
-
-    @GetMapping("/shared/top5")
-    @Operation(summary = "[공통] 조회수 상위 5개 포트폴리오 조회")
-    public ResponseEntity<List<PortfolioOverviewDTO.Response>> getTop5Portfolios() {
-        List<PortfolioOverviewDTO.Response> top5Portfolios = portfolioService.getTop5Portfolios();
-        return ResponseEntity.ok(top5Portfolios);
-    }
 
     @GetMapping("/weddingplanner/me")
     @Operation(summary = "[웨딩플래너] 내 포트폴리오 조회")
